Add fuzz CI implementation plan rename v1.rs to base.rs
